Moving all responsibility for running multiple steps into the StepList class.
