+++ lib/Build.js
@@ -4,6 +4,8 @@ var through2 = require('through2');
 var uuid = require('node-uuid');
 
 var _stream = Symbol('_stream');
+var _container = Symbol('_container');
+var _state = Symbol('_state');
 
 class Build extends events.EventEmitter {
 
@@ -18,51 +20,38 @@ class Build extends events.EventEmitter {
     super();
     options = options || {};
     this.id = options.id || uuid.v4();
-    this.emitErrors = options.emitErrors || false;
-    // The dockerode Container object on which this build should be performed.
+    // The probo Container object to run this build on.
     this.container = null;
-    // The array of provision steps that must be run in this build.
-    this.steps = options.steps || [];
-    // The step currently being run when iterating on steps.
-    this.position = 0;
+    // The step to run as this build.
+    this.step = options.step;
     this[_stream] = through2.obj();
-    this.endedStreams = [];
-    this.stepErrors = [];
-    // Used when iterating over the steps to determine the current step that we're running.
-    this.currentStepIndex = 0;
-    events.EventEmitter.call(this);
   }
 
-  setContainer(container) {
-    this.container = container;
-    container.build = this;
+  set container(container) {
+    this[_container] = container;
   }
 
-  addStep(step) {
-    this.steps.push(step);
-    step.build = this;
+  set state(state) {
+    this[_state] = state;
   }
+
+  get state() {
+    return this[_state];
+  }
+
 
   run(done) {
-    this.state = 'running';
-    if (this.steps.length === 0) {
-      return done();
-    }
-    this.runNextStep(function(error) {
-      if (done) {
-        return done(error);
-      }
+    var self = this;
+    // this.step.run(done);
+    self.emit('start', this);
+    self.state = 'running';
+    self.step.stream.pipe(self.stream);
+    self.step.run(function(error) {
+      self.stream.end();
+      self.emit('end', this);
+      self.state = error ? 'failed' : 'completed';
+      done(error);
     });
-  }
-
-  /**
-   * Get an object stream with the output of the steps running as a part of this build.
-   * TODO: make this a getter.
-   *
-   * @return {ReadableStream} - An object stream containing multiplexed stdout and stderr.
-   */
-  getStream() {
-    return this.stream;
   }
 
   get stream() {
@@ -74,83 +63,16 @@ class Build extends events.EventEmitter {
    *
    * @return {ReadableStream} - An text stream of multiplexed stdout and stderr from this build.
    */
-  getJSONStream() {
+  get jsonStream() {
     return this.stream.pipe(through2.obj(function(data, enc, cb) {
       cb(null, JSON.stringify(data) + '\n');
     }));
-  }
-
-  runNextStep(finalCallback) {
-    var self = this;
-    var step = this.steps[this.position];
-    self.emit('stepStart', step);
-    self.attachStream(step.getStream());
-    step.run(function(error) {
-      self.emit('stepEnd', step);
-      self.handleStepCompletion(error, finalCallback);
-    });
-  }
-
-  handleStepCompletion(error, finalCallback) {
-    var step = this.steps[this.position];
-    if (error) {
-      if (this.emitErrors) {
-        this.emit('error', error);
-      }
-      if (!step.optional) {
-        this.stepErrors.push(error);
-      }
-      if (!step.continueOnFailure) {
-        return finalCallback(error);
-      }
-    }
-    if (this.position + 1 === this.steps.length) {
-      this.state = 'completed';
-      if (this.stepErrors.length > 0) {
-        error = new Error('An optional step experienced an error.');
-        error.stepErrors = this.stepErrors;
-        if (!step.optional) {
-          this.state = 'failed';
-        }
-      }
-      return finalCallback(error);
-    }
-    this.position++;
-    this.runNextStep(finalCallback);
-  }
-
-  attachStream(stream) {
-    var self = this;
-    var itemDecorator = function(data, enc, cb) {
-      data.buildId = self.id;
-      this.push(data);
-      cb();
-    };
-    stream
-      .pipe(through2.obj(itemDecorator, self.streamEndHandler.bind(self, stream)))
-      .pipe(this.stream, {end: false});
-  }
-
-  streamEndHandler(stream) {
-    this.endedStreams.push(stream);
-    if (this.endedStreams.length === this.steps.length) {
-      this.stream.end();
-    }
-  }
-
-  handleError(error, done) {
-    if (this.emitErrors) {
-      this.emit('error', error);
-    }
-    return done(error);
   }
 
   save(done) {
   }
 }
 
-// TODO: Stop using *step* and *task* interchangably - it's steps in the yaml it
-// sould be steps everywhere.
 
 /*
