+++ lib/plugins/Step/AbstractStep.js
@@ -241,9 +241,7 @@ class AbstractStep extends events.EventEmitter {
       .pipe(self._getStringifyStream())
       .pipe(split2())
       .pipe(self.stdErrStream);
-
   }
-
 
   handleTimeout(done) {
     this.error = new Error(`${this.name} exited due to timeout.`);
@@ -251,16 +249,29 @@ class AbstractStep extends events.EventEmitter {
     this.container.stop(done);
   }
 
-  errorHandler(error, msg, done) {
-    msg = msg || error.message;
+  /**
+   * Generic error handler for errors that will properly emit events.
+   *
+   * @param {Error} error - The error event that need to be handled.
+   * @param {String} message - The error event that need to be handled.
+   * @param {Function} done - The error event that need to be handled.
+   */
+  errorHandler(error, message, done) {
+    if (typeof message === 'function') {
+      done = message;
+      message = null;
+    }
+    message = message || error.message;
 
     this.endTime = Date.now();
-    this.updateStatus({state: 'error', action: 'finished', description: msg});
+    this.updateStatus({state: 'error', action: 'finished', description: message});
 
-    this.log.error({err: error}, msg);
-    var wrappedError = new Error(msg);
+    this.log.error({err: error}, message);
+    var wrappedError = new Error(message);
     wrappedError.originalError = error;
-    return done(wrappedError);
+    if (done) {
+      done(wrappedError);
+    }
   }
 
   /**
